Add support for decoding optional values
Currently decoding an optional value entails using the `oneOf`
decoder, and passing in the decoder for the type you want decoded,
the `isUndefined(undefined)`, and `isNull(undefined)` to get an
optional value on an object.

The downside of doing this means that if you have an error on your
object, you lose the details of the error because the error becomes
`can't be decoded with any of the provided oneOf decoders` instead
of continuing the error trace down further down the structure to the
next oneOf decoders or the place where the decoding failed.